Created user creation functionality

+++ api/models/user_attributes.py
@@ -7,6 +7,6 @@ class UserAttributes(TimestampMixin, db.Model):
     __tablename__ = "user_attributes"
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"))
-    user_phone = db.Column(db.Integer, nullable=False, index=True)
+    user_phone = db.Column(db.BigInteger, nullable=False, index=True)
     field_name = db.Column(db.String(255))
     field_value = db.Column(db.String(255))
